index

+++ src/context/userContext.tsx
@@ -1,10 +1,16 @@
-import { createContext, useState, ReactNode, useEffect, useContext } from 'react';
+import { createContext, useState, ReactNode, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { CartContext } from './cartContext';
 import api from '../services/api';
 
 interface IDefaultProviderProps {
   children: ReactNode;
+}
+interface ICartProduct {
+  id: number;
+  name: string;
+  category?: string | undefined;
+  price?: number | undefined;
+  img: string;
 }
 
 interface IUserContext {
@@ -12,6 +18,8 @@ interface IUserContext {
   registerUser: (data: IRegisterUser) => void;
   user1: IUser | null;
   setUser: React.Dispatch<React.SetStateAction<IUser | null>>;
+  setProducts: React.Dispatch<React.SetStateAction<ICartProduct[]>>;
+  products: IProduct[];
 }
 interface IUser {
   email: string;
@@ -23,23 +31,27 @@ interface IRegisterUser {
   email: string;
   password: string;
 }
+interface IProduct {
+  id: number;
+  name: string;
+  category?: string | undefined;
+  price?: number | undefined;
+  img: string;
+}
 
 export const UserContext = createContext({} as IUserContext);
 
 export const UserProvider = ({ children }: IDefaultProviderProps) => {
   const [user1, setUser] = useState<IUser | null>(null);
-  const {getProducts} = useContext(CartContext)
   const navigate = useNavigate();
+  const [products, setProducts] = useState<IProduct[]>([]);
 
   const loginUser = async (formData: IUser) => {
-    // eslint-disable-next-line no-console
-    console.log(JSON.stringify(formData))
-    
     try {
-      const response = await api.post('/login', (formData));
+      const response = await api.post('/login', formData);
+      setUser(response.data.user);
       localStorage.setItem('@Token:', response.data.accessToken);
       localStorage.setItem('@USERID:', response.data.user.id);
-      setUser(response.data.user);
       navigate('/shop');
       getProducts();
     } catch (error) {
@@ -47,9 +59,9 @@ export const UserProvider = ({ children }: IDefaultProviderProps) => {
       alert(error);
     }
   };
-
   const registerUser = async (formData: IRegisterUser) => {
     try {
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const response = await api.post('/users', formData);
       navigate('/');
     } catch (error) {
@@ -57,28 +69,48 @@ export const UserProvider = ({ children }: IDefaultProviderProps) => {
       alert(error);
     }
   };
+
+  const getProducts = async () => {
+    try {
+      const response = await api.get('/products', {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem(`@Token:`)}`,
+        },
+      });
+      setProducts(response.data);
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.log('error');
+    }
+  };
+
   useEffect(() => {
-    async function autoLogin() {
-      try {
-        const response = await api.post(
-          `/login/${localStorage.getItem('@Token:')}`
-        );
-        setUser(response.data.user);
-        navigate(`/shop`);
-      } catch (error) {
-        // eslint-disable-next-line no-alert
-        alert('Faca login');
-        localStorage.clear();
-        navigate('/');
-      }
+    const token = localStorage.getItem('@Token:');
+    if (token) {
+      autoLogin();
+      getProducts();
     }
   }, []);
 
+  const autoLogin = async () => {
+    const token = localStorage.getItem('@Token:');
+    if (token) {
+      navigate('/shop');
+    } else {
+      // eslint-disable-next-line no-alert
+      alert('Faca login');
+      localStorage.clear();
+      navigate('/');
+    }
+  };
 
   return (
-    <UserContext.Provider value={{ loginUser, user1, setUser, registerUser }}>
+    <UserContext.Provider
+      value={{ loginUser, user1, setUser, registerUser, products, setProducts }}
+    >
       {children}
     </UserContext.Provider>
   );
 };
+
 export default UserProvider;
